Moved from ActiveMerchant::Billing::FirstData::* to ActiveMerchant::Billing::FirstData* namespace
