--- lib/active_merchant/billing/first_data.rb
@@ -1,12 +1,0 @@
-require "active_merchant/billing/first_data/gateway"
-require "active_merchant/billing/first_data/response"
-
-require "active_merchant/billing/first_data_response"
-require "active_merchant/billing/first_data_gateway"
-
-module ActiveMerchant #:nodoc:
-  module Billing #:nodoc:
-    module FirstData
-    end
-  end
-end
