+++ lib/active_merchant_first_data.rb
@@ -1,2 +1,3 @@
 require "active_merchant"
-require "active_merchant/billing/first_data"
+require "active_merchant/billing/first_data_response"
+require "active_merchant/billing/first_data_gateway"
